+++ autd3-firmware-emulator/src/error.rs
@@ -4,7 +4,7 @@
  * Created Date: 14/12/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 14/12/2023
+ * Last Modified: 20/12/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -18,4 +18,10 @@ use thiserror::Error;
 pub enum AUTDExtraError {
     #[error("The size of local_trans_pos is wrong.")]
     FPGALocalTransPos,
+}
+
+#[derive(Error, Debug)]
+pub enum AUTDFirmwareEmulatorError {
+    #[error("The input data is invalid.")]
+    InvalidDateTime,
 }